add cofirmacao ao excluir cliente

+++ src/Controladores/FXML_ClientesController.java
@@ -10,7 +10,10 @@ import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonBar;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -155,10 +158,26 @@ public class FXML_ClientesController implements Initializable {
 
     @FXML
     private void on_contExcluir(ActionEvent event) {
-
+        boolean r = false;
         try {
-            dao.remover(tabela_cliente.getSelectionModel().getSelectedItem().getId().getValue());
-            atualizarTabela();
+            Alert dialogo1 = new Alert(Alert.AlertType.CONFIRMATION);
+            dialogo1.setTitle("Atenção");
+            dialogo1.setHeaderText("Deseja realmente excluir o cliente " + tabela_cliente.getSelectionModel().getSelectedItem().getNome().getValue() + "?");
+            dialogo1.getButtonTypes().add(ButtonType.YES);
+            ButtonType btn = new ButtonType("Sim", ButtonBar.ButtonData.YES);
+            ButtonType btn1 = new ButtonType("Cancelar", ButtonBar.ButtonData.CANCEL_CLOSE);
+            dialogo1.getButtonTypes().setAll(btn, btn1);
+            r = dialogo1.showAndWait().get().equals(btn);
+            if (r) {
+                dao.remover(tabela_cliente.getSelectionModel().getSelectedItem().getId().getValue());
+                atualizarTabela();
+                Alert dialogo = new Alert(Alert.AlertType.INFORMATION);
+
+                dialogo.setTitle("Cliente excluido!");
+                dialogo.setHeaderText("Operaçao bem sucedida!");
+                dialogo.showAndWait();
+            }
+
         } catch (SQLException ex) {
             Logger.getLogger(FXML_ClientesController.class.getName()).log(Level.SEVERE, null, ex);
         }
